+++ extensions/markdown-slides/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Markdown Slides Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Commands to manage and preview presentations based on a markdown file
+- HTML export feature, with configuration options available when creating a new presentation
+- Quickly generate presentations using Raycast AI